+++ rotkehlchen/chain/ethereum/modules/nft/nfts.py
@@ -5,7 +5,6 @@ from typing import TYPE_CHECKING, Any, Optional
 from pysqlcipher3 import dbapi2 as sqlcipher
 
 from rotkehlchen.assets.asset import Asset
-from rotkehlchen.chain.evm.decoding.uniswap.v3.types import AddressToUniswapV3LPBalances
 from rotkehlchen.constants import ZERO
 from rotkehlchen.constants.assets import A_USD
 from rotkehlchen.db.filtering import NFTFilterQuery
@@ -216,11 +215,9 @@ class Nfts(EthereumModule, CacheableMixIn, LockableQueryMixIn):
     def query_balances(
             self,
             addresses: Sequence[ChecksumEvmAddress],
-            uniswap_nfts: AddressToUniswapV3LPBalances | None,
     ) -> None:
         """Queries NFT balances for the specified addresses and saves them to the db.
         Doesn't return anything. The actual opensea querying part is protected by a lock.
-        If `uniswap_nfts` is not None then the worth of the LPs are used as the value of the NFTs.
 
         May raise:
         - RemoteError
@@ -231,20 +228,12 @@ class Nfts(EthereumModule, CacheableMixIn, LockableQueryMixIn):
         queried_addresses = sorted(set(accounts.eth) & set(addresses))  # Sorting for consistency in tests  # noqa: E501
         nft_results, _ = self._get_all_nft_data(queried_addresses, ignore_cache=True)
         db_data: list[NFT_DB_WRITE_TUPLE] = []
-        # get uniswap v3 lp balances and update nfts that are LPs with their worth.
         for address, nfts in nft_results.items():
             for nft in nfts:
-                # get the lps for the address and check if the nft is a LP,
-                # then replace the worth with LP value.
-                uniswap_v3_lps = uniswap_nfts.get(address) if uniswap_nfts is not None else None
-                uniswap_v3_lp = next((entry for entry in uniswap_v3_lps if entry.nft_id == nft.token_identifier), None) if uniswap_v3_lps is not None else None  # noqa: E501
                 collection_name = nft.collection.name if nft.collection is not None else None
-                if uniswap_v3_lp is not None:
-                    db_data.append((nft.token_identifier, nft.name, str(uniswap_v3_lp.user_balance.usd_value), 'USD', False, address, True, nft.image_url, collection_name))  # noqa: E501
-                else:
-                    if collection_name == 'Uniswap V3 Positions':  # a uniswap v3 collection but is not detected in the balances... it is an exited position  # noqa: E501
-                        continue
-                    db_data.append((nft.token_identifier, nft.name, str(nft.price_in_asset), nft.price_asset.identifier, False, address, False, nft.image_url, collection_name))  # noqa: E501
+                if collection_name == 'Uniswap V3 Positions':  # skip all uniswap v3 positions - they are handled by UniswapV3Balances now.  # noqa: E501
+                    continue
+                db_data.append((nft.token_identifier, nft.name, str(nft.price_in_asset), nft.price_asset.identifier, False, address, False, nft.image_url, collection_name))  # noqa: E501
 
         # Update DB cache
         fresh_nfts_identifiers = [x[0] for x in db_data]